requirements.txt
pandas
numpy
scikit-learn
xgboost
lightgbm
nltk
seaborn
matplotlib
mlflow
dvc[s3]
PyYAML
fastapi
flask
wordcloud
-e .